backend/pyproject.toml
[tool.poetry]
authors = ["Lucía <lucia.blanco.rguez@gmail.com>"]
description = ""
name = "backend"
packages = [
  {include = "cleaner"},
  {include = "core"},
  {include = "hotels"},
  {include = "users"},
]
version = "0.1.0"

[tool.poetry.dependencies]
Django = "^4.2"
django-cors-headers = "^3.14.0"
django-extensions = "^3.2.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^4.8.0"
djoser = "^2.1.0"
drf-spectacular = "^0.26.1"
numpy = "^1.24.2"
pandas = "^2.0.0"
psycopg2-binary = "^2.9.6"
python = ">=3.8.11, <3.12"
scikit-learn = "^1.2.2"
scipy = "1.10.1"
tqdm = "^4.9.0"
pyparsing = "^3.1.0"
pydot = "^1.4.2"
selenium = "^4.10.0"
python-dotenv = "^1.0.0"
faker = "^23.1.0"
scrapy = "^2.11.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
